+++ packages/agent/src/actor.test.ts
@@ -6,7 +6,7 @@ import { CallRequest, SubmitRequestType, UnSigned } from './agent/http/types';
 import * as cbor from './cbor';
 import { requestIdOf } from './request_id';
 import * as pollingImport from './polling';
-import { Actor, ActorConfig } from './actor';
+import { ActorConfig } from './actor';
 
 const importActor = async (mockUpdatePolling?: () => void) => {
   jest.dontMock('./polling');
@@ -367,26 +367,6 @@ describe('makeActor', () => {
     expect(() => Actor.createActor(actorInterface, config)).toThrowError(
       'Canister ID is required, but received undefined instead. If you are using automatically generated declarations, this may be because your application is not setting the canister ID in process.env correctly.',
     );
-  });
-  
-  it.only('should allow for passing createCertificateOptions', async () => {
-    const idlFactory = ({ IDL }) =>
-      IDL.Service({
-        greet: IDL.Func([], [IDL.Text]),
-      });
-    const actor = Actor.createActor(idlFactory, {
-      canisterId: Principal.fromText('2chl6-4hpzw-vqaaa-aaaaa-c'),
-      agent: HttpAgent.createSync({
-        retryTimes: 0,
-        // Deliberately using a non-standard port to ensure the request fails
-        host: 'http://localhost:1111',
-      }),
-      createCertificateOptions: {
-        maxAgeInMinutes: 2,
-      },
-    });
-
-    expect(actor.greet).rejects.toThrowErrorMatchingInlineSnapshot();
   });
 });