+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@
 - feat: adds optional `rootKey` to `HttpAgentOptions` to allow for a custom root key to be used for verifying signatures from other networks
 - chore: npm audit bumping micromatch
 - feat: adds `maxAgeInMinutes` option to `HttpAgent` to allow for setting a maximum age for certificate verification
+- feat: adds `Certificate.validateOptions` method to check inputs when creating a certificate
 
 ### Changed
 - feat: replaces hdkey and bip32 implementations with `@scure/bip39` and `@scure/bip32` due to vulnerability and lack of maintenance for `elliptic`